Started on refactoring
