+++ importer-ui/src/features/main/hooks/useStepNavigation.ts
@@ -1,4 +1,8 @@
 import { useEffect, useState } from "react";
+import useStepper from "../../../components/Stepper/hooks/useStepper";
+import { Steps } from "../types";
+import useModifiedSteps from "./useModifiedSteps";
+import useMutableLocalStorage from "./useMutableLocalStorage";
 
 export const StepEnum = {
   Upload: 0,
@@ -32,11 +36,22 @@ const calculateCurrentStep = (currentStep: number, skipHeader: boolean) => {
   }
 };
 
-function useStepNavigation(initialStep: number, skipHeader: boolean) {
+export const stepsConfig = [
+  { label: "Upload", id: Steps.Upload },
+  { label: "Select Header", id: Steps.RowSelection },
+  { label: "Map Columns", id: Steps.MapColumns },
+  { label: "Review", id: Steps.Review },
+];
+
+function useStepNavigation(initialStep: number, skipHeader: boolean, importerId: string) {
+  const steps = useModifiedSteps(stepsConfig, skipHeader);
+  const stepper = useStepper(steps, 0);
+  const [storageStep, setStorageStep] = useMutableLocalStorage(`tf_steps_${importerId}`, "");
   const [currentStep, setCurrentStep] = useState(initialStep);
 
   const setStep = (newStep: number) => {
     setCurrentStep(newStep);
+    setStorageStep(newStep);
   };
 
   useEffect(() => {
@@ -44,8 +59,11 @@ function useStepNavigation(initialStep: number, skipHeader: boolean) {
   }, [skipHeader]);
 
   return {
-    currentStep,
+    currentStep: storageStep ? parseInt(storageStep) : currentStep,
     setStep,
+    stepper,
+    setStorageStep,
+    storageStep,
   };
 }
